Adds items endpoints
Builds out all endpoints for items
as described by spec. All local
tests are currently passing.
Serializer generated for item
which excludes return of timestamps
as per spec. Spec harness is being
slightly wonky still. Given the errors
I think tackling the import of the data
next would be a great idea, so we won't be
aiming in the dark.